+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes for this project will be documented in this file.
 
+## 0.0.41 - 2018-06-28
+### Changed
+- when adding research to REAL content, resources are now added to a collection that corresponds to the media
+- removed selection from other models -- won't be used the same way as a REAL model
+
 ## 0.0.40 - 2018-06-27
 ### Added
 - selection prop to OLesson and OSermon for adding individual resources